==> 05_richness-vegetation.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Associations between bird species richness, detections and Vegetation data

Previously, we calculated species richness and number of detections for every 16-min duration across sites and treatment types. Here, we ask if there is a statistically significant association between bird species richness and/or detections and vegetation measurements

Install required libraries
```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(vegan)
library(ggplot2)
library(scico)
library(psych)

# Source any custom/other internal functions necessary for analysis
source("code/01_internal-functions.R")
```


Load the vegetation data
```{r}
veg <- read.csv("data/2020-vegetation-data.csv")
veg$Site_ID <- str_remove(veg$Site_ID,"_")

# Subset data to include only sites that are present in our study
# To do so, we load the 2020-annotation-data and extract the unique sites
data <- read.csv("data/2020-summer-annotation-working document.csv")

# Split the file names into 4 columns : Site, Date, Time and Splits   
data <- separate(data, col = Filename, into = c("Site", "Date", "Time", "Splits"), sep = "_")
sites_needed <- data.frame(unique(data$Site))
names(sites_needed) <- "Site_ID"

# Obtain a subset of the data which have the sites visited (Note: 44/69 sites were visited)
veg <- merge(veg, sites_needed, by.x = "Site_ID", by.y="Site_ID")

# Load species-trait data to essentially run GLMs
trait_dat <- read.csv("data/species-trait-dat.csv")
```


Carrying out exploratory analysis and preparing a dataframe for further steps
```{r}
# Counting number of tree species and unique species per plot
treerich <- veg %>% 
  group_by(Site_ID) %>% 
  summarise (count = n(), richness = n_distinct(tree_species))

# Calculate average tree height across each unique site
treeheight <- veg %>% 
  drop_na(height) %>% 
  group_by(Site_ID) %>% 
  summarise(height = mean(height))

# Calculate basal area and left join with other data
basal_area <- veg %>% 
  mutate(basal_sum = rowSums(veg[,c(5:15)]^2)/(4*pi)) %>% group_by(Site_ID, Site_type) %>% 
  summarise(basal_area = sum(basal_sum)) 

# Calculate average canopy height
canopy_height <- veg %>%
  group_by(Site_ID) %>%
  summarise(canopy_cover = mean(Canopy_cover))

# Calculate average leaf litter
leaf_litter <- veg %>%
  group_by(Site_ID) %>%
  summarise(leaf_litter = mean(Leaf_litter))

# Calculate average vertical stratification
vert_strat <- veg %>%
  group_by(Site_ID) %>%
  summarise(vert_strat = mean(Foliage_score))

# Creating a final dataframe for further analysis
all_veg <- basal_area %>% 
  left_join(treeheight) %>%
  left_join(treerich) %>%
  left_join(canopy_height) %>%
  left_join(leaf_litter) %>%
  left_join(vert_strat)
```


Running a PCA of the vegetation data
```{r}
# Check for correlations among vegetation predictors
pairs.panels(all_veg[,3:9])

# The above panel suggests that richness is highly correlated with a number of predictors including canopy cover, count and basal area. We will calculate a PCA and keep the top two explanatory axes

veg_pca <- prcomp(all_veg[, 3:9], scale=TRUE, center = TRUE, retx=TRUE)
summary(veg_pca)

# The proportion of variance explained by the first two axes account for ~72.67%

# Plotting the first two axes
plot_data <- data.frame('Site_type' = all_veg$Site_type, veg_pca$x[,1:2]) # the first two components are selected 
pca_plot <- ggplot(data = plot_data) + 
       geom_point(aes(x = PC1, y = PC2, col = Site_type), size=2) + 
  scale_color_manual(values = scico(3, begin = 0.3, palette = "vik")) + theme_bw()
```


Running a GLMM that compares avian species and vegetation specific
```{r}
# To run a GLMM, we need to first extract necessary predictors that will be use in the mixed-effects model. 

# 1. Calculate the overall number of detections for each site across X days of data (translates to 16min to 48min of data per site; each detection corresponding to a temporal unit of 10seconds). Here, we 

nDetections_site_date <- data %>%
  group_by(Site, Restoration.Type, Date) %>%
  summarise_at(.vars = vars(c("IP":"CR")),.funs = sum) %>% 
  transform() %>% replace(is.na(.), 0)

# 2. Combine the nDetections and trait based data to obtain a dataframe for jackknife estimates

nDetections_trait <- nDetections_site_date %>%
  pivot_longer(cols=IP:CR, names_to="Species_Code", values_to="count") %>%
  left_join(.,trait_dat, by=c("Species_Code"="species_annotation_codes")) %>%
  dplyr::select(-c(scientific_name)) %>%
  mutate(forRichness = case_when(count>0 ~ 1,count==0 ~ 0)) %>%
  rename(., nDetections = count)

nSites_Days <- data %>%
  dplyr::select(Site, Date)%>%
  distinct() %>% arrange(Site) %>% count(Site)




 filt_data %>% 
  group_by(Site_type) %>%
  summarise (
    detections = n(), 
    species = n_distinct(Species), 
    abundance = sum(Number),
    visits = n_distinct(Site_ID, Date))



 
 

```
