==> 11_glmms-acoustic-space-use-vegPca.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Generalized linear mixed modeling (acoustic space use and vegetation data)

In this script, we run generalized linear models to test the association between acoustic space use values and restoration type. In addition, we run generalized linear mixed models to test associations between acoustic space use and habitat (vegetation structure) using site-pair name (actively restored and naturally regenerating were specified to be paired) and repeat visits as random effects.  

Install required libraries
```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(vegan)
library(ggplot2)
library(scico)
library(psych)
library(rcompanion)
library(multcomp)
library(lme4)
library(ggpubr)
library(sjPlot)
library(RColorBrewer)

# Source any custom/other internal functions necessary for analysis
source("code/01_internal-functions.R")
```

Load necessary data for statistical modeling
```{r}
# load list of sites
sites <- read.csv("data/list-of-sites.csv") %>% 
  dplyr::select("Site.code","Restoration.type") %>%
  filter(Site.code != "OLCAP5B")

# load the entire asu data across all sites and days computed
sitebyDayAsu <- read.csv("data/site-by-day-asu.csv") 

# separate by Site and Date
sitebyDayAsu <- separate(sitebyDayAsu, col = Site_Day, into = c("Site", "Date"), sep = "_") 

# Add restoration type column to the space use data
sitebyDayAsu <- left_join(sitebyDayAsu, sites, by=c("Site"="Site.code"))

# scale values per site/date for comparison between sites and treatment types
sitebyDayAsu <- sitebyDayAsu %>%
  group_by(Site, Date, Restoration.type) %>%
  mutate(f.cont.scaled = range01(f.cont))

# computing total number of days for which space use has been computed for each site (some sites have very few days. Eg. OLV110R, else rest have 5 days of audio data each)
nSites_Days <- sitebyDayAsu %>%
  dplyr::select(Site, Date, Restoration.type) %>%
  distinct() %>%
  group_by(Site) %>%
  count()

# Let's look at data by restoration type
# This suggests that we have more data for benchmark sites relative to the other two treatment types
nDays_siteType <- sitebyDayAsu %>%
  dplyr::select(Site, Date, Restoration.type) %>%
  distinct() %>%
  group_by(Restoration.type) %>%
  count()
  
# Separate analysis: space use during certain times of day alone
# let's look only at dawn chorus to explore what the data looks like
dawn <- c("05:00-06:00","06:00-07:00","07:00-08:00","08:00-09:00",
          "09:00-10:00")
dawnChorusAsu <- sitebyDayAsu %>%
  filter(time_of_day %in% dawn)

# Separate analysis: space use during certain times of day alone
# let's look only at dusk chorus to explore what the data looks like
dusk <- c("16:00-17:00","17:00-18:00","18:00-19:00","19:00-20:00")
duskChorusAsu <- sitebyDayAsu %>%
  filter(time_of_day %in% dusk)

# Prepare data for statistical modeling
# Calculating total space use across all frequency bins and times of day: 128*24 for each site-day combination

totSpaceUse <- sitebyDayAsu %>%
  group_by(Site, Date, Restoration.type) %>%
  summarise(totSpaceuse = sum(f.cont.scaled)) %>%
  group_by(Site) %>% 
  mutate(visit = row_number()) %>%
  mutate(siteCode = str_extract(Site, pattern = "\\w+\\d+")) %>%
  mutate(siteCode = factor(siteCode))

# Calculating total space use for dawn chorus alone for each site-day combination

totSpaceUseDawn <- dawnChorusAsu %>%
  group_by(Site, Date, Restoration.type) %>%
  summarise(totSpaceuse = sum(f.cont.scaled)) %>% 
  group_by(Site) %>% 
  mutate(visit = row_number()) %>%
  mutate(siteCode = str_extract(Site, pattern = "\\w+\\d+")) %>%
  mutate(siteCode = factor(siteCode))

# Calculating total space use for dusk chorus alone for each site-date combination

totSpaceUseDusk <- duskChorusAsu %>%
  group_by(Site, Date, Restoration.type) %>%
  summarise(totSpaceuse = sum(f.cont.scaled)) %>% 
  group_by(Site) %>% 
  mutate(visit = row_number()) %>%
  mutate(siteCode = str_extract(Site, pattern = "\\w+\\d+")) %>%
  mutate(siteCode = factor(siteCode))

# Load data from previous scripts for use in a GLM
vegData <-  read.csv("data/summaryVeg.csv") %>%
  filter(!str_detect(Site_ID, 'OLCAP5B'))
vegPcaScores <- read.csv("data/pcaVeg.csv") %>%
  filter(!str_detect(Site_ID, 'OLCAP5B'))
```


Getting data ready in a format for linear modeling
```{r}
# overall space use
modelDataAll <- vegPcaScores %>% 
  rename(Site = Site_ID) %>%
  rename(Restoration.type = Site_type) %>%
  mutate(across(Restoration.type, factor))  %>%
  full_join(totSpaceUse, by=c("Site","Restoration.type")) %>%
  mutate("roundSpaceuse" = round(totSpaceuse))

# dawn chorus space use
modelDataDawn <- vegPcaScores %>% 
  rename(Site = Site_ID) %>%
  rename(Restoration.type = Site_type) %>%
  mutate(across(Restoration.type, factor))  %>%
  full_join(totSpaceUseDawn, by=c("Site","Restoration.type")) %>%
  mutate("roundSpaceuse" = round(totSpaceuse))

# dusk chorus space use
modelDataDusk <- vegPcaScores %>% 
  rename(Site = Site_ID) %>%
  rename(Restoration.type = Site_type) %>%
  mutate(across(Restoration.type, factor))  %>%
  full_join(totSpaceUseDusk, by=c("Site","Restoration.type")) %>%
  mutate("roundSpaceuse" = round(totSpaceuse))
```

Running the generalized linear mixed models

We now run generalized linear mixed models (GLMM) assuming gaussian errors to examine the effects of restoration type (benchmark, actively restored and passively restored) on acoustic space use, followed by TukeyHSD multiple comparisons tests of means. 

```{r}
# overall space use
glmm_allRestoration <- glmer(roundSpaceuse ~ Restoration.type + (1|siteCode) + (1|visit), data = modelDataAll, family = gaussian(link="identity"))

summary(glmm_allRestoration)
tukey_glmmAllRestoration <- summary(glht(glmm_allRestoration, linfct=mcp(Restoration.type ="Tukey")))
cld(tukey_glmmAllRestoration)

# The above result suggests a significant difference in acoustic space use between BM-AR and BM-NR sites, but no difference between AR and NR sites. 

# dawn chorus
glmm_dawnRestoration <- glmer(roundSpaceuse ~ Restoration.type + (1|siteCode) + (1|visit), data = modelDataDawn, family = gaussian(link="identity"))

summary(glmm_dawnRestoration)
tukey_glmmdawnRestoration <- summary(glht(glmm_dawnRestoration, linfct=mcp(Restoration.type ="Tukey")))
cld(tukey_glmmdawnRestoration)

# When we look at only the dawn chorus data, there is no significant difference in acoustic space use across all treatment types highlighting similar levels of vocalizing biodiversity. 

# dusk chorus
glmm_duskRestoration <- glmer(roundSpaceuse ~ Restoration.type + (1|siteCode) + (1|visit), data = modelDataDusk, family=gaussian(link="identity"))

summary(glmm_duskRestoration)
tukey_glmmduskRestoration <- summary(glht(glmm_duskRestoration, linfct=mcp(Restoration.type ="Tukey")))
cld(tukey_glmmduskRestoration)

# For dusk chorus data, the above result suggests a significant difference between BM-AR and BM-NR sites, but no difference between AR and NR sites. 

# Plotting the above results

# reordering factors for plotting
modelDataAll$Restoration.type <- factor(modelDataAll$Restoration.type, levels = c("Benchmark", "Active", "Passive"))

fig_glmm_allRest <- ggplot(modelDataAll, aes(Restoration.type, totSpaceuse, fill = Restoration.type)) + 
  geom_boxplot(alpha = 0.7) + 
  scale_fill_scico_d(palette = "roma") +
    theme_bw() +
    labs(x="\nTreatment Type", 
       y="Acoustic Space Use\n") +
  scale_x_discrete(labels = c('BM','AR','NR')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
        legend.position = "none")

ggsave(fig_glmm_allRest, filename = "figs/fig_glmm_overallAcousticSpace.png", width=12, height=7, device = png(), units="in", dpi = 300);dev.off()

# reordering factors for plotting
modelDataDawn$Restoration.type <- factor(modelDataDawn$Restoration.type, levels = c("Benchmark", "Active", "Passive"))

fig_glmm_dawnRest <- ggplot(modelDataDawn, aes(Restoration.type, totSpaceuse, group = Restoration.type, fill = Restoration.type)) + 
 geom_boxplot(alpha = 0.7) + 
  scale_fill_scico_d(palette = "roma") +
    theme_bw() +
    labs(x="\nTreatment Type", 
       y="Acoustic Space Use\n") +
  scale_x_discrete(labels = c('BM','AR','NR')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
        legend.position = "none")

ggsave(fig_glmm_dawnRest, filename = "figs/fig_glmm_dawnAcousticSpaceUse.png", width=12, height=7, device = png(), units="in", dpi = 300);dev.off()

# reordering factors for plotting
modelDataDusk$Restoration.type <- factor(modelDataDusk$Restoration.type, levels = c("Benchmark", "Active", "Passive"))

fig_glmm_duskRest <- ggplot(modelDataDusk, aes(Restoration.type, totSpaceuse, group = Restoration.type, fill = Restoration.type)) + 
  geom_boxplot(alpha = 0.7) + 
  scale_fill_scico_d(palette = "roma") +
    theme_bw() +
    labs(x="\nTreatment Type", 
       y="Acoustic Space Use\n") +
  scale_x_discrete(labels = c('BM','AR','NR')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
        legend.position = "none")
  
ggsave(fig_glmm_duskRest, filename = "figs/fig_glmm_duskAcousticSpaceUse.png", width=12, height=7, device = png(), units="in", dpi = 300);dev.off()
```

![**Overall acoustic space use across treatment types.**
Analysis of ~4,128 hours of acoustic data across treatment types (for the 43 sites) revealed significant differences in ASU between treatment types (Tukey HSD P < 0.05). ASU was significantly different across BM-AR, BM-NR and AR-NR sites, with the highest estimate in BM sites (mean ± SD: 413 ± 103), followed by AR sites (mean ± SD: 188 ± 78) and NR sites (mean ± SD: 182 ± 66). Mean and standard deviation were based on pooled values of ASU across 24 hours for every site-day combination. In the above figure, BM = undisturbed benchmark rainforest sites, AR = Actively restored forest sites and NR = Naturally regenerating forest sites.](figs/fig_glmm_overallAcousticSpace.png)  

We now run generalized linear mixed models (GLMM) assuming gaussian errors and using log link functions to examine the effects of habitat (vegetation structure) on acoustic space use.

```{r}
# overall space use
glmm_allVeg<- glmer(roundSpaceuse ~ PC1 + PC2 + (1|siteCode) + (1|visit), data = modelDataAll, family = gaussian(link="identity"))

summary(glmm_allVeg)
plot_model(glmm_allVeg, type="pred", terms=c("PC1","PC2"))
report::report(glmm_allVeg)

# There is a significant negative effect of PC1 on acoustic space use

# dawn chorus
glmm_dawnVeg<- glmer(roundSpaceuse ~ PC1 + PC2 + (1|siteCode) + (1|visit), data = modelDataDawn, family = gaussian(link="identity"))

summary(glmm_dawnVeg)
plot_model(glmm_dawnVeg, type="pred", terms=c("PC2","PC1"))
report::report(glmm_dawnVeg)

# There is a significant negative effect of PC1 on dawn acoustic space use

# dawn chorus
glmm_duskVeg<- glmer(roundSpaceuse ~ PC1 + PC2 + (1|siteCode) + (1|visit), data = modelDataDusk, family = gaussian(link="identity"))

summary(glmm_duskVeg)
plot_model(glmm_duskVeg, type="pred", terms=c("PC1","PC2"))
report::report(glmm_duskVeg)

# There is a significant negative effect of PC1 on dusk acoustic space use
```

Creating figure for publication
```{r}
# Patchworking nmds and glmm for acoustic space use to create part a and b for Fig. 4

library(patchwork)

fig_glmm_ordinations <-
  wrap_plots(fig_glmm_allRest, fig_nmds,
    design = "AABBBB"
  ) +
  plot_annotation(
    tag_levels = "a",
    tag_prefix = "(",
    tag_suffix = ")"
  )

# Expand the width to avoid compression
ggsave(fig_glmm_ordinations, filename = "figs/fig04.png", width=20, height=7,device = png(), units="in", dpi = 300); dev.off()
```
