==> 13_glms-acoustic-space-use-richness-plantingYear.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Generalized linear modeling (acoustic space use and richness, and time since restoration)

In this script, we run generalized linear models to test the association between acoustic space use values and species richness, vocal activity (total calling rate at each site, which is a proxy for the total number of vocal detections/unit of recording time) and time since restoration. 

Install required libraries
```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(vegan)
library(ggplot2)
library(scico)
library(psych)
library(rcompanion)
library(multcomp)
library(lme4)
library(ggpubr)
library(sjPlot)

# Source any custom/other internal functions necessary for analysis
source("code/01_internal-functions.R")
```

Load necessary data for statistical modeling
```{r}
# load list of sites
sites <- read.csv("data/list-of-sites.csv") %>% 
  dplyr::select("Site.code","Restoration.type") %>%
  filter(Site.code != "OLCAP5B")

# loading jacknife scores 
jackAll <- read.csv("data/jackAll.csv")
jack_rainForest <- read.csv("data/jackRainforest.csv")
jack_openCountry <- read.csv("data/jackOpencountry.csv")

# Load vegetation data from previous scripts 
vegData <-  read.csv("data/summaryVeg.csv") %>%
  filter(!str_detect(Site_ID, 'OLCAP5B'))
vegPcaScores <- read.csv("data/pcaVeg.csv") %>%
  filter(!str_detect(Site_ID, 'OLCAP5B'))

# Attach the annotation data
datSubset <- read.csv("data/datSubset.csv")

# Calculate the overall number of detections for each site where each temporal duration chosen is a 10s clip
nDetections <- datSubset %>%
  group_by(Site, Restoration.type) %>%
  transform() %>% replace(is.na(.), 0) %>% 
  summarise_at(.vars = vars(c("IP":"HSWP")),.funs = sum) 

# total number of detections at each site
totDetections <-  nDetections %>%
   rowwise() %>% 
   summarise(totDetections = sum(c_across(IP:HSWP)))

# load the entire asu data across all sites and days computed
sitebyDayAsu <- read.csv("data/site-by-day-asu.csv") 

# separate by Site and Date
sitebyDayAsu <- separate(sitebyDayAsu, col = Site_Day, into = c("Site", "Date"), sep = "_") 

# Add restoration type column to the space use data
sitebyDayAsu <- left_join(sitebyDayAsu, sites, by=c("Site"="Site.code"))

# scale values per site/date for comparison between sites and treatment types
sitebyDayAsu <- sitebyDayAsu %>%
  group_by(Site, Date, Restoration.type) %>%
  mutate(f.cont.scaled = range01(f.cont))

# computing total number of days for which space use has been computed for each site (some sites have very few days. Eg. OLV110R, else rest have 5 days of audio data each)
nSites_Days <- sitebyDayAsu %>%
  dplyr::select(Site, Date, Restoration.type) %>%
  distinct() %>%
  group_by(Site) %>%
  count()

# Let's look at data by restoration type
# This suggests that we have more data for benchmark sites relative to the other two treatment types
nDays_siteType <- sitebyDayAsu %>%
  dplyr::select(Site, Date, Restoration.type) %>%
  distinct() %>%
  group_by(Restoration.type) %>%
  count()
  
# Prepare data for statistical modeling
# Calculating total space use across all frequency bins and times of day: 128*24 for each site-day combination

totSpaceUse <- sitebyDayAsu %>%
  group_by(Site, Date, Restoration.type) %>%
  summarise(totSpaceuse = sum(f.cont.scaled)) %>%
  group_by(Site) %>% 
  mutate(visit = row_number()) %>%
  mutate(siteCode = str_extract(Site, pattern = "\\w+\\d+")) %>%
  mutate(siteCode = factor(siteCode)) %>%
  full_join(vegData, by=c("Site"="Site_ID"))

```


Getting data ready in a format for linear modeling
```{r}
# overall space use
modelDataAll <- vegPcaScores %>% 
  rename(Site = Site_ID) %>%
  rename(Restoration.type = Site_type) %>%
  mutate(across(Restoration.type, factor))  %>%
  full_join(totSpaceUse, by=c("Site","Restoration.type")) %>%
  mutate("roundSpaceuse" = round(totSpaceuse)) %>%
  full_join(jackAll, by=c("Site","Restoration.type")) %>% 
  mutate ("roundjk" = round(jack1)) %>%
  full_join(vegData, by=c("Site"="Site_ID")) %>%
  full_join(totDetections, by=c("Site"="Site"))

# overall space use and rainforest birds
modelDataRain <- vegPcaScores %>% 
  rename(Site = Site_ID) %>%
  rename(Restoration.type = Site_type) %>%
  mutate(across(Restoration.type, factor))  %>%
  full_join(totSpaceUse, by=c("Site","Restoration.type")) %>%
  mutate("roundSpaceuse" = round(totSpaceuse)) %>%
  full_join(jack_rainForest, by=c("Site","Restoration.type")) %>% 
  mutate ("roundjk" = round(jack1)) %>%
  full_join(vegData, by=c("Site"="Site_ID"))

# overall space use and open country birds
modelDataOpen <- vegPcaScores %>% 
  rename(Site = Site_ID) %>%
  rename(Restoration.type = Site_type) %>%
  mutate(across(Restoration.type, factor))  %>%
  full_join(totSpaceUse, by=c("Site","Restoration.type")) %>%
  mutate("roundSpaceuse" = round(totSpaceuse)) %>%
  full_join(jack_openCountry, by=c("Site","Restoration.type")) %>% 
  mutate ("roundjk" = round(jack1)) %>%
  full_join(vegData, by=c("Site"="Site_ID"))

```


Let's first explore richness (as measured through first order jacknife scores and test for associations with space use)
```{r}
# running generalized linear models
glmm_overallRichnessSpace <- glmer(roundSpaceuse ~ roundjk + (1|siteCode) + (1|visit), data = modelDataAll, family = gaussian(link="identity"))

summary(glmm_overallRichnessSpace)
plot_model(glmm_overallRichnessSpace, type="pred")
report::report(glmm_overallRichnessSpace)

# The results suggest a statistically non-significant and postive association between overall space use and first-order jacknife scores

# Rainforest species
glmm_RainRichnessSpace <- glmer(roundSpaceuse ~ roundjk + (1|siteCode) + (1|visit), data = modelDataRain, family = gaussian(link="identity"))

summary(glmm_RainRichnessSpace)
plot_model(glmm_RainRichnessSpace, type="pred")
report::report(glmm_RainRichnessSpace)

# for rainforest species,  The effect of roundjk is statistically significant and positive (beta = 5.30, 95% CI [1.98, 8.62], t(205) = 3.15, p = 0.002; Std. beta = 0.24, 95% CI [0.09, 0.40])

# open country species
glmm_OpenRichnessSpace <- glmer(roundSpaceuse ~ roundjk + (1|siteCode) + (1|visit), data = modelDataOpen, family = gaussian(link="identity"))

summary(glmm_OpenRichnessSpace)
plot_model(glmm_OpenRichnessSpace, type="pred")
report::report(glmm_OpenRichnessSpace)

# For open country birds, The effect of roundjk is statistically non-significant and negative (beta = -1.56, 95% CI [-4.29, 1.18], t(205) = -1.12, p = 0.263; Std. beta = -0.07, 95% CI [-0.19, 0.05])
```


Let's look at year since restoration and ask if space use has a significant association with time
```{r}
allRestored <- totSpaceUse %>%
  filter(Restoration.type=="Active") %>%
  mutate(yearSinceRestoration = (2022-plantingYear))

glmmAllYear <- glmer(totSpaceuse ~ yearSinceRestoration + (1|visit), data = allRestored, family = gaussian(link="identity"))
summary(glmmAllYear)

plot_model(glmmAllYear, type="pred")
report::report(glmmAllYear)

# The effect of yearSinceRestoration is statistically significant and negative (beta = -10.07, 95% CI [-16.66, -3.47], t(61) = -3.05, p = 0.003; Std. beta = -0.35, 95% CI [-0.58, -0.12])

```
